fix windows filename case renaming issue
