+++ packages/delegate/src/index.ts
@@ -3,7 +3,7 @@ export { createRequestFromInfo, createRequest } from './createRequest';
 export { defaultMergedResolver } from './defaultMergedResolver';
 export { createMergedResolver } from './createMergedResolver';
 export { handleResult } from './results/handleResult';
-export { Subschema, isSubschema, isSubschemaConfig, getSubschema } from './Subschemas';
+export { Subschema, isSubschema, isSubschemaConfig, getSubschema } from './Subschema';
 
 export * from './transforms';
